Implementação do handshake criptográfico Ref #3 e reorganização do projeto em pastas.
