+++ .gitignore
@@ -1,3 +1,4 @@
 __pycache__
 chat_history.db
-*.pem
+*_private_key.pem
+*_public_key.pem
